refactor(portfolio-contract): Better separate construction of network graphs vs. LP solver models
