+++ packages/portfolio-contract/tools/plan-solve.ts
@@ -1,11 +1,11 @@
-import type { Amount } from '@agoric/ertp';
-import { AmountMath } from '@agoric/ertp';
-import type { NatAmount } from '@agoric/ertp/src/types.js';
-
-import { Fail } from '@endo/errors';
 import makeHighsSolver from 'highs';
 import jsLPSolver from 'javascript-lp-solver';
 import type { IModel, IModelVariableConstraint } from 'javascript-lp-solver';
+
+import { Fail } from '@endo/errors';
+
+import { AmountMath } from '@agoric/ertp';
+import type { Amount, NatAmount } from '@agoric/ertp/src/types.js';
 import {
   makeTracer,
   naturalCompare,
@@ -14,19 +14,16 @@ import {
   provideLazyMap,
   typedEntries,
 } from '@agoric/internal';
-import { AxelarChain } from '@agoric/portfolio-api/src/constants.js';
-import { PoolPlaces, type PoolKey } from '../src/type-guards.js';
+import type { AxelarChain } from '@agoric/portfolio-api/src/constants.js';
+
 import type { AssetPlaceRef, MovementDesc } from '../src/type-guards-steps.js';
-import type {
-  FeeMode,
-  NetworkSpec,
-  TransferProtocol,
-} from './network/network-spec.js';
-import { makeGraphFromDefinition } from './network/buildGraph.js';
 import {
   preflightValidateNetworkPlan,
   formatInfeasibleDiagnostics,
 } from './graph-diagnose.js';
+import { chainOf, makeGraphFromDefinition } from './network/buildGraph.js';
+import type { FlowEdge, RebalanceGraph } from './network/buildGraph.js';
+import type { NetworkSpec } from './network/network-spec.js';
 
 const highsSolverP = makeHighsSolver();
 
@@ -42,42 +39,6 @@ const replaceOrInit = <K, V>(
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const trace = makeTracer('solve');
-
-// ----------------------------------- Types -----------------------------------
-
-/**
- * A chain hub node id looks like '@agoric', '@noble', '@Arbitrum', etc.
- * Leaf nodes: PoolKey, '<Cash>', '<Deposit>', '+agoric'
- */
-
-/** Internal edge representation */
-export interface FlowEdge {
-  id: string;
-  src: AssetPlaceRef;
-  dest: AssetPlaceRef;
-  capacity: number; // numeric for LP; derived from bigint
-  variableFee: number; // cost coefficient per unit flow in basis points
-  fixedFee?: number; // optional fixed cost (cheapest mode)
-  timeFixed?: number; // optional time cost (fastest mode)
-  via?: TransferProtocol; // annotation (e.g. 'intra-chain', 'cctp', etc.)
-  feeMode?: FeeMode;
-}
-
-/** Node supply: positive => must send out; negative => must receive */
-export interface SupplyMap {
-  [node: string]: number;
-}
-
-/** Graph structure */
-export interface RebalanceGraph {
-  nodes: Set<AssetPlaceRef>;
-  edges: FlowEdge[];
-  supplies: SupplyMap;
-  brand: Amount['brand'];
-  feeBrand: Amount['brand'];
-  /** When true, print extra diagnostics on solver failure */
-  debug?: boolean;
-}
 
 /** Mode of optimization */
 export type RebalanceMode = 'cheapest' | 'fastest';
@@ -111,8 +72,6 @@ type GasEstimator = {
   getReturnFeeEstimate: (chainName: AxelarChain) => Promise<bigint>;
 };
 
-// --- keep existing type declarations above ---
-
 /**
  * -----------------------------------------------------------------------------
  * Scaling Rules (model domain normalization)
@@ -125,95 +84,6 @@ type GasEstimator = {
  * -----------------------------------------------------------------------------
  */
 const FLOW_EPS = 1e-6;
-
-/**
- * Build base graph with:
- * - Hub nodes for each chain discovered in placeRefs (auto-added as '@chain')
- * - Leaf nodes for each placeRef (except hubs already formatted)
- * - Intra-chain bidirectional edges leaf <-> hub (variableFee=1, timeFixed=1)
- * - Supplies = current - target; if target missing, assume unchanged (target=current)
- */
-export const buildBaseGraph = (
-  placeRefs: AssetPlaceRef[],
-  current: Partial<Record<AssetPlaceRef, NatAmount>>,
-  target: Partial<Record<AssetPlaceRef, NatAmount>>,
-  brand: Amount['brand'],
-  feeBrand: Amount['brand'],
-): RebalanceGraph => {
-  const nodes = new Set<AssetPlaceRef>();
-  const edges: FlowEdge[] = [];
-  const supplies: SupplyMap = {};
-
-  // Collect chains needed
-  for (const ref of placeRefs) {
-    nodes.add(ref);
-    const chain = chainOf(ref);
-    nodes.add(`@${chain}` as AssetPlaceRef);
-  }
-
-  // Build supplies (signed deltas)
-  for (const node of nodes) {
-    const currentVal = current[node]?.value ?? 0n;
-    const targetSpecified = Object.prototype.hasOwnProperty.call(target, node);
-    const targetVal = targetSpecified ? target[node]!.value : currentVal; // unchanged if unspecified
-    const delta = currentVal - targetVal;
-    // NOTE: Number(bigint) loses precision beyond MAX_SAFE_INTEGER (2^53-1)
-    // For USDC amounts (6 decimals), the largest realistic value would be trillions
-    // of dollars, which should be well within safe integer range.
-    if (delta !== 0n) supplies[node] = Number(delta);
-  }
-
-  // Ensure intra-chain edges (leaf <-> hub)
-  let eid = 0;
-  const vf = 1; // direct variable fee per unit
-  const tf = 1; // time cost unit (seconds or abstract)
-  for (const node of nodes) {
-    const chainName = chainOf(node);
-    const hub = `@${chainName}` as AssetPlaceRef;
-    // Skip hubs and agoric-local places (which are connected unidirectionally).
-    if (node === hub || hub === '@agoric') continue;
-
-    const chainIsEvm = Object.keys(AxelarChain).includes(chainName);
-    const base: Omit<FlowEdge, 'src' | 'dest' | 'id'> = {
-      capacity: 1e15,
-      variableFee: vf,
-      fixedFee: 0,
-      timeFixed: tf,
-      via: 'local',
-    };
-
-    edges.push({
-      // eslint-disable-next-line no-plusplus
-      id: `e${eid++}`,
-      src: node,
-      dest: hub,
-      ...base,
-      ...(chainIsEvm ? { feeMode: 'poolToEvm' } : {}),
-    });
-
-    // Skip @agoric → +agoric edge
-    if (node === '+agoric') continue;
-
-    edges.push({
-      // eslint-disable-next-line no-plusplus
-      id: `e${eid++}`,
-      src: hub,
-      dest: node,
-      ...base,
-      ...(chainIsEvm ? { feeMode: 'evmToPool' } : {}),
-    });
-  }
-
-  // Return mutable graph (do NOT harden so we can add inter-chain links later)
-  return {
-    nodes,
-    edges,
-    supplies,
-    brand,
-    feeBrand,
-    debug: false,
-  } as RebalanceGraph;
-};
 
 // ------------------------------ Model Building -------------------------------
 
@@ -638,23 +508,6 @@ export const planRebalanceFlow = async (opts: {
   const { flows, detail } = result;
   const steps = await rebalanceMinCostFlowSteps(flows, graph, gasEstimator);
   return harden({ graph, model, flows, steps, detail });
-};
-
-const chainOf = (id: AssetPlaceRef): string => {
-  if (id.startsWith('@')) return id.slice(1);
-  if (id === '<Cash>' || id === '<Deposit>' || id === '+agoric')
-    return 'agoric';
-  if (id in PoolPlaces) {
-    const pk = id as PoolKey;
-    return PoolPlaces[pk].chainName;
-  }
-  // Fallback: syntactic pool id like "Protocol_Chain" => chain
-  // This enables base graph edges for pools even if not listed in PoolPlaces
-  const m = /^([A-Za-z0-9]+)_([A-Za-z0-9-]+)$/.exec(id);
-  if (m) {
-    return m[2];
-  }
-  throw Fail`Cannot determine chain for ${id}`;
 };
 
 // ---------------------------- Example (commented) ----------------------------
